Refactor: Did some stuff.
Refs: BLAIS5-3120
