+++ src/components/SurveyDateForm.tsx
@@ -1,7 +1,6 @@
 import React, { ReactElement } from "react";
 import { StyledForm } from "blaise-design-system-react-components";
-import { formatISODate } from "../utilities/DateFormatter";
-import { SurveyField } from "./SurveyField";
+import { DateField, SurveyField } from "./FormFields";
 
 interface Props {
     onSubmitFunction: (values: any, setSubmitting: (isSubmitting: boolean) => void) => void;
@@ -11,11 +10,7 @@ const SurveyDateForm = ({ onSubmitFunction }: Props): ReactElement => {
 
     const fields = [
         SurveyField("undefined"),
-        {
-            name: "Date",
-            type: "date",
-            initial_value: formatISODate(new Date()),
-        }
+        DateField()
     ];
 
     return (
