+++ src/components/SurveyInterviewerStartDateEndDateForm.tsx
@@ -1,7 +1,6 @@
 import React, { ReactElement } from "react";
 import { StyledForm } from "blaise-design-system-react-components";
-import { formatISODate } from "../utilities/DateFormatter";
-import { SurveyField } from "./SurveyField";
+import { EndDateField, InterviewerField, StartDateField, SurveyField } from "./FormFields";
 
 interface Props {
     interviewer: string | undefined
@@ -19,53 +18,15 @@ const SurveyInterviewerStartDateEndDateForm = ({
     onSubmitFunction
 }: Props): ReactElement => {
 
-    const validateInterviewer = (value: string) => {
-        let error;
-        if (value === "" || value === undefined) {
-            error = "Enter a interviewer ID";
-        }
-        return error;
-    };
-
-    const validateDate = (value: string) => {
-        let error;
-        if (value === "" || value === undefined) {
-            error = "Enter a date";
-            return error;
-        }
-
-        if (value.match(/^[0-9]{4}-[0-9]{2}-[0-9]{2}$/) === null) {
-            error = "Enter a valid date";
-        }
-        return error;
-    };
-
     const fields = [
         SurveyField(surveyTLA),
-        {
-            name: "Interviewer ID",
-            type: "text",
-            initial_value: interviewer,
-            validate: validateInterviewer
-        },
-        {
-            name: "Start date",
-            type: "date",
-            initial_value: formatISODate(startDate),
-            validate: validateDate
-        },
-        {
-            name: "End date",
-            type: "date",
-            initial_value: formatISODate(endDate),
-            validate: validateDate
-        }
+        InterviewerField(interviewer),
+        StartDateField(startDate),
+        EndDateField(endDate),
     ];
 
     return (
-        <>
-            <StyledForm fields={fields} onSubmitFunction={onSubmitFunction} submitLabel={"Next"} />
-        </>
+        <><StyledForm fields={fields} onSubmitFunction={onSubmitFunction} submitLabel={"Next"}/></>
     );
 };
